requirements.txt
Django==1.7
Pillow==2.5.1
PyYAML==3.11
amqp==1.4.5
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.13
django-bootstrap3==4.9.0
django-configurations==0.8
django-debug-toolbar==1.2.1
django-extensions==1.3.10
django-mptt==0.6.1
django-waffle==0.10
django-braces==1.4.0
kombu==3.0.21
psycopg2==2.5.3
pytz==2014.4
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
